Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -12,14 +12,14 @@
 class Solution {
 public:
     long long  res=0;
-    int  pathSum(TreeNode* root, long long  targetSum) {
-        unordered_map<long long ,long long > dict; // sum-> cnt; 
+    int  pathSum(TreeNode* root, int  targetSum) {
+        unordered_map<long long ,int > dict; // sum-> cnt; 
         // dict[0]=1;
         dfs(root,targetSum,0,dict);
         return res;
     }
 
-    void dfs(TreeNode* node,long long & targetSum,long long  sum,unordered_map<long long ,long long > dict){
+    void dfs(TreeNode* node,int & targetSum,long long  sum,unordered_map<long long ,int > dict){
         if(!node) return ;
         sum+=node->val;
         if(sum==targetSum) res++;
